pyproject.toml
[tool.poetry]
authors = ["Hironori Yamamoto <mr.nikoru918@gmail.com>"]
description = ""
name = "chatgpt-playground"
packages = [{include = "chatgpt_playground"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.10"

chromadb = "^0.3.17"
click = "^8.1.3"
google-cloud-bigquery = "^3.9.0"
langchain = "^0.0.126"
openai = "^0.27.2"
pydantic = "^1.10.7"
structlog = "^22.3.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
invoke = "^2.0.0"
isort = "^5.10.1"
mypy = "^0.982"
pytest = "^7.1.3"
pytest-asyncio = "^0.20.3"
pytest-cov = "^4.0.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry.scripts]
gpt = 'chatgpt_playground.cli:cli'

[tool.mypy]
exclude = [
  "tasks.py",
]
plugins = [
  "pydantic.mypy",
]

check_untyped_defs = true
disallow_any_decorated = false
disallow_any_generics = false
disallow_any_unimported = false
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
disallow_untyped_defs = true
ignore_errors = false
ignore_missing_imports = false
no_implicit_optional = true
no_implicit_reexport = true
python_version = "3.10"
show_error_codes = true
strict_equality = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = ["pandas.*", "async_asgi_testclient.*"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
known_first_party = ["chatgpt_playground"]
line_length = 120
multi_line_output = 3
use_parentheses = true

[tool.coverage.run]
source = ["chatgpt_playground"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
